realized TestModule for s_calculate funct | add 5 Test Cases

+++ lab08/prj/TestDriver/ModulesStriuk/modulesstriuk.cpp
@@ -2,7 +2,7 @@
 
 const double PI = 3.14159265;
 
-double s_calculation(int x, int y, int z)
+float s_calculation(float x, float y, float z)
 {
     return log(x - y) + sqrt((PI * pow(x, 2)) / x + (z / (2 * pow(y, 2))));
 }
